Rich Text Elements

RichTextElements

Description

(simple but) rich text formatting tools

julia> import BibTeXFormat: RichText, Tag, render_as, add_period, capitalize

julia> t = RichText("this ", "is a ", Tag("em", "very"), RichText(" rich", " text"));

julia> render_as(t,"LaTex")
"this is a \\emph{very} rich text"

julia> convert(String,t)
"this is a very rich text"

julia> t = add_period(capitalize(t));

julia> render_as(t,"latex")
"This is a \\emph{very} rich text."

Types

BibTeXFormat.HRef

BibTeXFormat.Protected

BibTeXFormat.RichString

BibTeXFormat.RichString

Functions

Base.:+

Base.:==

Base.UTF8proc.isalpha

Base.UTF8proc.isalpha

Base.UTF8proc.isalpha

Base.contains

Base.endswith

Base.endswith

Base.getindex

Base.join

Base.length

Base.lowercase

Base.split

Base.split

Base.startswith

Base.startswith

Base.uppercase

BibTeXFormat.add_period

BibTeXFormat.append

BibTeXFormat.append

BibTeXFormat.capfirst

BibTeXFormat.capitalize

BibTeXFormat.create_similar

BibTeXFormat.initialize_parts

BibTeXFormat.merge_similar

BibTeXFormat.slice_beginning

BibTeXFormat.slice_end

BibTeXFormat.typeinfo

Reference

BibTeXFormat.HRef — Type.

A HRef represends a hyperlink:

julia> href = HRef("http://ctan.org/", "CTAN");

julia> print(render_as(href,"html"))
<a href="http://ctan.org/">CTAN</a>

julia> print(render_as(href, "latex"))
\href{http://ctan.org/}{CTAN}

julia> href = HRef(String("http://ctan.org/"), String("http://ctan.org/"));

julia> print(render_as(href,"latex"))
\url{http://ctan.org/}

source

BibTeXFormat.Protected — Type.

A Protected represents a "protected" piece of text.

Protected.lowercase, Protected.uppercase, Protected.capitalize, and Protected.capitalize() are no-ops and just return the Protected struct itself.

split never splits the text. It always returns a one-element list containing the Protected struct itself.

In LaTeX output, Protected is {surrounded by braces}. HTML and plain text backends just output the text as-is.

´´´jldoctest julia> text = Protected("The CTAN archive"); julia> lowercase(text) BibTeXFormat.Protected(Any[BibTeXFormat.RichString("The CTAN archive")], 16, "")

julia> print(split(text)) BibTeXFormat.Protected[BibTeXFormat.Protected(Any[BibTeXFormat.RichString("The CTAN archive")], 16, "")]

julia> print(render_as(text, "latex")) {The CTAN archive}

julia> print(render_as(text,"html")) <span class="bibtex-protected">The CTAN archive</span>

´´´

source

BibTeXFormat.RichString — Type.

A RichString is a wrapper for a plain Julia string.

julia> print(render_as(RichString("Crime & Punishment"),"text"))
Crime & Punishment
julia> print(render_as(RichString("Crime & Punishment"),"html"))
Crime &amp; Punishment

source

BibTeXFormat.RichString — Method.

All arguments must be plain unicode strings. Arguments are concatenated together.

julia> print(convert(String,RichString("November", ", ", "December", ".")))
November, December.

source

Base.:+ — Method.

function +(b::BaseText, other)

Concatenate this Text with another Text or string.

julia> a = RichText("Longcat is ") + Tag("em", "long")
RichText("Longcat is ", Tag("em", "long"))

source

Base.:== — Method.

Compare two RichString objects.

source

Base.UTF8proc.isalpha — Method.

Return True if all characters in the string are alphabetic and there is at least one character, False otherwise.

source

Base.UTF8proc.isalpha — Method.

A TextSymbol is not alfanumeric. Returns false

source

Base.UTF8proc.isalpha — Method.

Return True if all characters in the string are alphabetic and there is at least one character, False otherwise.

source

Base.contains — Method.

$value in text$ returns $True$ if any part of the $text$ contains the substring $value$:

julia> contains(RichText("Long cat!"),"Long cat")
true

Substrings splitted across multiple text parts are not matched:

julia> contains(RichText(Tag("em", "Long"), "cat!"),"Long cat")
false

source

Base.endswith — Method.

Return True if the string ends with the specified suffix, otherwise return False.

suffix can also be a tuple of suffixes to look for. return self.value.endswith(text)

source

Base.endswith — Method.

Return True if the text ends with the given suffix.

julia> endswith(RichText("Longcat!"),"cat!")
true

Suffixes split across multiple parts are not matched:

julia> endswith(RichText("Long", Tag("em", "cat"), "!"),"cat!")
false

source

Base.getindex — Method.

Slicing and extracting characters works like with regular strings, formatting is preserved.

julia> RichText("Longcat is ", Tag("em", "looooooong!"))[1:15]
RichText("Longcat is ", Tag("em", "looo"))

julia> RichText("Longcat is ", Tag("em", "looooooong!"))[end]
RichText(Tag("em", "!"))

source

Base.join — Method.

function join(self::T, parts) where T<:BaseText

Join a list using this text (like join)

julia> letters = ["a", "b", "c"];

julia> print(convert(String,join(RichString("-"),letters)))
a-b-c

source

Base.length — Method.

$lenght(text)$ returns the number of characters in the text, ignoring the markup:

julia> length(RichText("Long cat"))
8
julia> length(RichText(Tag("em", "Long"), " cat"))
8
julia> length(RichText(HRef("http://example.com/", "Long"), " cat"))
8

source

Base.lowercase — Method.

Convert rich text to lowercasecase.

julia> lowercase(RichText(Tag("em", "Long cat")))
RichText(Tag("em", "long cat"))

source

Base.split — Function.

Split

source

Base.split — Method.

julia> print(split(RichText("a + b")))
Any[RichText("a"), RichText("+"), RichText("b")]
julia> print(split(RichText("a, b"), ", "))
Any[RichText("a"), RichText("b")]

source

Base.startswith — Method.

Return True if string starts with the prefix, otherwise return False.

prefix can also be a tuple of suffixes to look for.

source

Base.startswith — Method.

Return True if the text starts with the given prefix.

julia> startswith(RichText("Longcat!"),"Longcat")
true

Prefixes split across multiple parts are not matched:

julia> startswith(RichText(Tag("em", "Long"), "cat!"),"Longcat")
false

source

Base.uppercase — Method.

Convert rich text to uppsercase.

julia> uppercase(RichText(Tag("em", "Long cat")))
RichText(Tag("em", "LONG CAT"))

source

BibTeXFormat.add_period — Function.

function add_period(self::BaseText, period=".")

Add a period to the end of text, if the last character is not ".", "!" or "?".

julia> text = RichText("That's all, folks");

julia> print(convert(String,add_period(text)))
That's all, folks.

julia> text = RichText("That's all, folks!");

julia> print(convert(String,add_period(text)))
That's all, folks!

source

BibTeXFormat.append — Method.

function append(self::BaseText, text)

Append text to the end of this text.

Normally, this is the same as concatenating texts with +, but for tags and similar objects the appended text is placed _inside_ the tag.

julia> text = Tag("em", "Look here");

julia> print(render_as(text + "!","html"))
<em>Look here</em>!

julia> print(render_as(append(text,"!"),"html"))
<em>Look here!</em>

source

BibTeXFormat.append — Method.

Append text to the end of this text.

For Tags, HRefs, etc. the appended text is placed inside the tag.

julia> text = Tag("strong", "Chuck Norris");

julia> print(render_as(text +  " wins!","html"))
<strong>Chuck Norris</strong> wins!
julia> print(render_as(append(text," wins!"),"html"))
<strong>Chuck Norris wins!</strong>

source

BibTeXFormat.capfirst — Method.

function capfirst(self::BaseText)

Capitalize the first letter of the text.

julia> capfirst(RichText(Tag("em", "long Cat")))
RichText(Tag("em", "Long Cat"))

source

BibTeXFormat.capitalize — Method.

function capitalize(self::BaseText)

Capitalize the first letter of the text and lowercasecase the rest.

julia> capitalize(RichText(Tag("em", "LONG CAT")))
RichText(Tag("em", "Long cat"))

source

BibTeXFormat.create_similar — Method.

function create_similar(self::T, parts) where T<:MultiPartText

Create a new text object of the same type with the same parameters, with different text content.

julia> text = Tag("strong", "Bananas!");

julia> create_similar(text,["Apples!"])
Tag("strong", "Apples!")

source

BibTeXFormat.initialize_parts — Method.

Initialize the parts. A MutliPartText elements must have the following members

type atype <: MultiPartText
	parts
	length
	info
end

Empty parts are ignored:

julia> RichText() == RichText("") == RichText("", "", "")
true
julia> RichText("Word", "") == RichText("Word")
true

Text() objects are unpacked and their children are included directly:

julia> RichText(RichText("Multi", " "), Tag("em", "part"), RichText(" ", RichText("text!")))
RichText("Multi ", Tag("em", "part"), " text!")

julia> Tag("strong", RichText("Multi", " "), Tag("em", "part"), RichText(" ", "text!"))
Tag("strong", "Multi ", Tag("em", "part"), " text!")

Similar objects are merged together:

julia> RichText("Multi", Tag("em", "part"), RichText(Tag("em", " ", "text!")))
RichText("Multi", Tag("em", "part text!"))

julia> RichText("Please ", HRef("/", "click"), HRef("/", " here"), ".")
RichText("Please ", HRef("/", "click here"), ".")

source

BibTeXFormat.merge_similar — Method.

Merge adjacent text objects with the same type and parameters together.

julia> parts = [Tag("em", "Breaking"), Tag("em", " "), Tag("em", "news!")];

julia> print(merge_similar(parts))
Any[Tag("em", "Breaking news!")]

source

BibTeXFormat.slice_beginning — Method.

Return a text consistng of the first slice_length characters of this text (with formatting preserved).

source

BibTeXFormat.slice_end — Method.

Return a text consistng of the last slice_length characters of this text (with formatting preserved).

source

BibTeXFormat.typeinfo — Method.

function  typeinfo(self::T) where T<:MultiPartText

Return the type and the parameters used to create this text object.

julia> text = Tag("strong", "Heavy rain!");

julia> typeinfo(text) == ("BibTeXFormat.Tag", BibTeXFormat.Tag, "strong")
true

source