Utilities

Utilities

Types

BibTeXFormat.BibTeXString

BibTeXFormat.abbreviate

BibTeXFormat.bibtex_abbreviate

BibTeXFormat.bibtex_first_letter

BibTeXFormat.bibtex_len

BibTeXFormat.bibtex_purify

BibTeXFormat.bibtex_substring

BibTeXFormat.change_case

BibTeXFormat.latex_parse

BibTeXFormat.scan_bibtex_string

BibTeXFormat.split_keep_separator

BibTeXFormat.split_name_list

BibTeXFormat.split_tex_string

Reference

BibTeXFormat.BibTeXString — Type.

julia> import BibTeXFormat: BibTeXString

julia> a = BibTeXString("{aaaa{bbbb{cccc{dddd}}}ffff}");

julia> convert(String,a ) == "{aaaa{bbbb{cccc{dddd}}}ffff}"
true

source

BibTeXFormat.abbreviate — Function.

Abbreviate the given text.

julia> import BibTeXFormat.abbreviate

julia> abbreviate("Name")
"N."

julia> abbreviate("Some words")
"S. w."

julia> abbreviate("First-Second")
"F.-S."

source

BibTeXFormat.bibtex_abbreviate — Function.

function bibtex_abbreviate(string, delimiter=None, separator='-')

Abbreviate string.

julia> import BibTeXFormat: bibtex_abbreviate

julia> print(bibtex_abbreviate("Andrew Blake"))
A
julia> print(bibtex_abbreviate("Jean-Pierre"))
J.-P
julia> print(bibtex_abbreviate("Jean--Pierre"))
J.-P

source

BibTeXFormat.bibtex_first_letter — Method.

function  bibtex_first_letter(string)

Return the first letter or special character of the string.

julia> import BibTeXFormat: bibtex_first_letter

julia> print(bibtex_first_letter("Andrew Blake"))
A
julia> print(bibtex_first_letter("{Andrew} Blake"))
A
julia> print(bibtex_first_letter("1Andrew"))
A
julia> print(bibtex_first_letter("{\\TeX} markup"))
{\TeX}
julia> print(bibtex_first_letter(""))

julia> print(bibtex_first_letter("123 123 123 {}"))

julia> print(bibtex_first_letter("\\LaTeX Project Team"))
L

source

BibTeXFormat.bibtex_len — Method.

Return the number of characters in the string.

function bibtex_len(string)

Braces are ignored. "Special characters" are ignored. A "special character" is a substring at brace level 1, if the first character after the opening brace is a backslash, like in "de la Vall{\'e}e Poussin".

julia> import BibTeXFormat: bibtex_len

julia> print(bibtex_len("de la Vall{\\'e}e Poussin"))
20
julia> print(bibtex_len("de la Vall{e}e Poussin"))
20
julia> print(bibtex_len("de la Vallee Poussin"))
20
julia> print(bibtex_len("\\ABC 123"))
8
julia> print(bibtex_len("{\\abc}"))
1
julia> print(bibtex_len("{\\abc"))
1
julia> print(bibtex_len("}\\abc"))
4
julia> print(bibtex_len("\\abc}"))
4
julia> print(bibtex_len("\\abc{"))
4
julia> print(bibtex_len("level 0 {1 {2}}"))
11
julia> print(bibtex_len("level 0 {\\1 {2}}"))
9
julia> print(bibtex_len("level 0 {1 {\\2}}"))
12

source

BibTeXFormat.bibtex_purify — Method.

function bibtex_purify(str)

Strip special characters from the string.

julia> import BibTeXFormat: bibtex_purify

julia> print(bibtex_purify("Abc 1234"))
Abc 1234
julia> print(bibtex_purify("Abc  1234"))
Abc  1234
julia> print(bibtex_purify("Abc-Def"))
Abc Def
julia> print(bibtex_purify("Abc-~-Def"))
Abc   Def
julia> print(bibtex_purify("{XXX YYY}"))
XXX YYY
julia> print(bibtex_purify("{XXX {YYY}}"))
XXX YYY
julia> print(bibtex_purify("XXX {\\YYY} XXX"))
XXX  XXX
julia> print(bibtex_purify("{XXX {\\YYY} XXX}"))
XXX YYY XXX
julia> print(bibtex_purify("\\abc def"))
abc def
julia> print(bibtex_purify("a@#\$@#\$b@#\$@#\$c"))
abc
julia> print(bibtex_purify("{\\noopsort{1973b}}1973"))
1973b1973
julia> print(bibtex_purify("{sort{1973b}}1973"))
sort1973b1973
julia> print(bibtex_purify("{sort{\\abc1973b}}1973"))
sortabc1973b1973
julia> print(bibtex_purify("{\\noopsort{1973a}}{\\switchargs{--90}{1968}}"))
1973a901968

source

BibTeXFormat.bibtex_substring — Method.

function bibtex_substring(string, start, length)

Return a substring of the given length, starting from the given position.

start and length are 1-based. If start is < 0, it is counted from the end of the string. If start is 0, an empty string is returned.

julia> import BibTeXFormat: bibtex_substring

julia> print(bibtex_substring("abcdef", 1, 3))
abc
julia> print(bibtex_substring("abcdef", 2, 3))
bcd
julia> print(bibtex_substring("abcdef", 2, 1000))
bcdef
julia> print(bibtex_substring("abcdef", 0, 1000))

julia> print(bibtex_substring("abcdef", -1, 1))
f
julia> print(bibtex_substring("abcdef", -1, 2))
ef
julia> print(bibtex_substring("abcdef", -2, 3))
cde
julia> print(bibtex_substring("abcdef", -2, 1000))
abcde

source

BibTeXFormat.change_case — Method.

function change_case(string, mode)

julia> import BibTeXFormat: change_case

julia> print(change_case("aBcD", 'l'))
abcd
julia> print(change_case("aBcD", 'u'))
ABCD
julia> print(change_case("ABcD", 't'))
Abcd
julia> change_case("The {\\TeX book \\noop}", 'u')
"THE {\\TeX BOOK \\noop}"

julia> change_case("And Now: BOOO!!!", 't')
"And now: Booo!!!"

julia> change_case("And {Now: BOOO!!!}", 't')
"And {Now: BOOO!!!}"

julia> change_case("And {Now: {BOOO}!!!}", 'l')
"and {Now: {BOOO}!!!}"

julia> change_case("And {\\Now: BOOO!!!}", 't')
"And {\\Now: booo!!!}"

julia> change_case("And {\\Now: {BOOO}!!!}", 'l')
"and {\\Now: {booo}!!!}"

julia> change_case("{\\TeX\\ and databases\\Dash\\TeX DBI}", 't')
"{\\TeX\\ and databases\\Dash\\TeX DBI}"

source

BibTeXFormat.scan_bibtex_string — Method.

Yield (char, brace_level) tuples.

"Special characters", as in bibtex_len, are treated as a single character

source

BibTeXFormat.split_keep_separator — Function.

Split a text keep the separators

julia> import BibTeXFormat.split_keep_separator

julia> split_keep_separator("Some words-words")
5-element Array{Any,1}:
 "Some"
 ' '
 "words"
 '-'
 "words"

source

BibTeXFormat.split_name_list — Method.

Split a list of names, separated by ' and '.

julia> import BibTeXFormat.split_name_list

julia> split_name_list("Johnson and Peterson")
2-element Array{String,1}:
 "Johnson"
 "Peterson"

julia> split_name_list("Johnson AND Peterson")
2-element Array{String,1}:
 "Johnson"
 "Peterson"

julia> split_name_list("Johnson AnD Peterson")
2-element Array{String,1}:
 "Johnson"
 "Peterson"

julia> split_name_list("Armand and Peterson")
2-element Array{String,1}:
 "Armand"
 "Peterson"

julia> split_name_list("Armand and anderssen")
2-element Array{String,1}:
 "Armand"
 "anderssen"

julia> split_name_list("{Armand and Anderssen}")
1-element Array{String,1}:
 "{Armand and Anderssen}"

julia> split_name_list("What a Strange{ }and Bizzare Name! and Peterson")
2-element Array{String,1}:
 "What a Strange{ }and Bizzare Name!"
 "Peterson"

julia> split_name_list("What a Strange and{ }Bizzare Name! and Peterson")
2-element Array{String,1}:
 "What a Strange and{ }Bizzare Name!"
 "Peterson"

source

BibTeXFormat.split_tex_string — Function.

Split a string using the given separator (regexp).

Everything at brace level > 0 is ignored. Separators at the edges of the string are ignored.

julia> import BibTeXFormat.split_tex_string

julia> split_tex_string("")
0-element Array{Any,1}

julia> split_tex_string("     ")
0-element Array{String,1}

julia> split_tex_string("   ", " ", strip=false, filter_empty=false)
2-element Array{Any,1}:
 " "
 " "

julia> split_tex_string(".a.b.c.", r"\.")
3-element Array{String,1}:
 ".a"
 "b"
 "c."

julia> split_tex_string(".a.b.c.{d.}.", r"\.")
4-element Array{String,1}:
 ".a"
 "b"
 "c"
 "{d.}."

julia> split_tex_string("Matsui      Fuuka")
2-element Array{String,1}:
 "Matsui"
 "Fuuka"

julia> split_tex_string("{Matsui      Fuuka}")
1-element Array{String,1}:
 "{Matsui      Fuuka}"

julia> split_tex_string(r"Matsui Fuuka")
2-element Array{String,1}:
 "Matsui"
 "Fuuka"

julia> split_tex_string("{Matsui Fuuka}")
1-element Array{String,1}:
 "{Matsui Fuuka}"

julia> split_tex_string("a")
1-element Array{String,1}:
 "a"

julia> split_tex_string("on a")
2-element Array{String,1}:
 "on"
 "a"

source

BibTeXFormat.latex_parse — Method.

from nose.tools import assert_raises

LaTeXParser('abc').parse()

Text('abc')

LaTeXParser('abc{def}').parse()

Text('abc', Protected('def'))

LaTeXParser('abc{def {xyz}} !').parse()

Text('abc', Protected('def ', Protected('xyz')), ' !')

assert_raises(PybtexSyntaxError, LaTeXParser('abc{def}}').parse) assert_raises(PybtexSyntaxError, LaTeXParser('abc{def}{').parse)

source