Style

Style

BST

using BibTeX
using BibTeXFormat
bibliography = Bibliography(readstring(joinpath(Pkg.dir("BibTeXFormat"), "test/Clustering.bib")))
style        = BST.parse_file(joinpath(Pkg.dir("BibTeXFormat"),"test/format/apacite.bst"))
formatted_entries = format_entries(style, bibliography)
HTMLoutput        = write_to_string( HTMLBackend(),formatted_entries )

BibTeXFormat.AlphaStyle — Constant.

Alpha Style

Config(label_style = AlphaLabelStyle(),sorting_style = AuthorYearTitleSortingStyle())

source

BibTeXFormat.format_entries — Method.

function format_entries(b::T, entries::Dict) where T <: BaseStyle

Format a Dict of entries with a given style b::T where T <: BaseStyle

using BibTeX
using BibTeXFormat
bibliography      = Bibliography(readstring("test/Clustering.bib"))
formatted_entries = format_entries(AlphaStyle,bibliography)

source

BibTeXFormat.PlainAlphaStyle — Constant.

PlainAlphaStyle

source

BibTeXFormat.UNSRTAlphaStyle — Constant.

UNSRTAlphaStyle

Config(label_style = AlphaLabelStyle())

source

BibTeXFormat.expand_wildcard_citations — Method.

Expand wildcard citations (citation{*} in .aux file).

julia> using BibTeX

julia> import BibTeXFormat: expand_wildcard_citations

julia> data = Bibliography("", Dict{String,Citation}("uno"=>Citation{:article}(),"dos"=>Citation{:article}(),"tres"=>Citation{:article}(),	"cuatro"=>Citation{:article}()));

julia> expand_wildcard_citations(data, [])
0-element Array{Any,1}

julia> print(expand_wildcard_citations(data, ["*"]))
Any["tres", "dos", "uno", "cuatro"]
julia> print(expand_wildcard_citations(data, ["uno", "*"]))
Any["uno", "tres", "dos", "cuatro"]
julia> print(expand_wildcard_citations(data, ["dos", "*"]))
Any["dos", "tres", "uno", "cuatro"]
julia> print(expand_wildcard_citations(data, ["*", "uno"]))
Any["tres", "dos", "uno", "cuatro"]
julia> print(expand_wildcard_citations(data, ["*", "DOS"]))
Any["tres", "dos", "uno", "cuatro"]

source

BibTeXFormat.format_bibliography — Method.

function format_bibliography(self::T, bib_data, citations=nothing) where T<:BaseStyle

Format bibliography entries with the given keys and return a FormattedBibliography object.

Params:

self::T where T<:BaseStyle. The style

bib_data A :py:class:pybtex.database.BibliographyData object.

param citations: A list of citation keys.

using BibTeX
using BibTeXFormat
bibliography      = Bibliography(readstring("test/Clustering.bib"))

formatted_entries = format_entries(AlphaStyle,bibliography)

source

BibTeXFormat.format_entry — Method.

function format_entry(b::T, label, entry) where T <: BaseStyle

Format an entry with a given style b::T where T <: BaseStyle

source

BibTeXFormat.get_crossreferenced_citations — Method.

Get cititations not cited explicitly but referenced by other citations.

julia> using BibTeX

julia> import BibTeXFormat: get_crossreferenced_citations

julia> data = Bibliography("", Dict{String,Citation}("main_article"=>Citation{:article}(Dict("crossref"=>"xrefd_article")),"xrefd_article"=>Citation{:article}()));

julia> print(get_crossreferenced_citations(data, [], min_crossrefs=1))
Any[]
julia> print(get_crossreferenced_citations(data, ["main_article"], min_crossrefs=1))
Any["xrefd_article"]
julia> print(get_crossreferenced_citations(data,["Main_article"], min_crossrefs=1))
Any["xrefd_article"]
julia> print(get_crossreferenced_citations(data, ["main_article"], min_crossrefs=2))
Any[]
julia> print(get_crossreferenced_citations(data, ["xrefd_arcicle"], min_crossrefs=1))
WARNING: bad cross-reference: entry "xrefd_arcicle" refers to
                entry "nothing" which does not exist.
Any[]

source