Template Engine

Template Engine

BibTeXFormat.TemplateEngine.format — Method.

A convenience function to be used instead of format_data when no data is needed.

source

BibTeXFormat.TemplateEngine.format_data — Method.

Format the given data into a piece of richtext.Text

source